Rename legacy query engines
